move batchable to kv module
